@@ -15,14 +15,14 @@
     "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
+    "@types/node": "20.16.8",
+    "@types/react": "18.3.9",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.4.20",
     "eslint": "^8",
     "eslint-config-next": "14.2.11",
     "postcss": "^8.4.45",
     "tailwindcss": "^3.4.11",
-    "typescript": "^5"
+    "typescript": "5.6.2"
   }
 }
